Refactor notifiaction_broadcaster.dart as notification_notifier.dart.
Delete old saved copies of notifiaction_broadcaster.dart and settings_page_contents.dart.

Refactor comments.
